requirements.txt
gradio
byaldi
qwen-vl-utils
numpy==1.24.4
Pillow==10.3.0
torch
torchvision
git+https://github.com/huggingface/transformers.git
accelerate